// sdks/typescript/src/generated/RedactionInfo.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RedactedEntity } from "./RedactedEntity.js";
import type { RedactionMode } from "./RedactionMode.js";
import type { RedactionStatus } from "./RedactionStatus.js";

export type RedactionInfo = { mode: RedactionMode, status: RedactionStatus, entities: Array<RedactedEntity>, input_redacted: boolean, proposed_output_redacted: boolean, context_redacted: boolean, };